+++ pyupgw/client.py
@@ -11,7 +11,7 @@ from collections.abc import Callable, Iterable, Mapping
 import aiohttp
 from dict_deep import deep_get
 
-from ._api import AwsApi, ServiceApi
+from ._api import AwsApi, ServiceApi, is_authentication_error
 from ._mqtt import IotShadowMqtt
 from .errors import AuthenticationError, ClientError
 from .models import (
@@ -429,12 +429,7 @@ async def create_api(username: str, password: str) -> AwsApi:
     try:
         await asyncio.to_thread(aws.authenticate, password)
     except Exception as ex:
-        # This is a convoluted way of figuring out if the exception is
-        # NotAuthorizedException. botocore generates exception types on the fly,
-        # so I can't just import the correct exception type and except it...
-        if "NotAuthorized" in str(
-            deep_get(ex, ["__class__", "__name__"], getter=getattr)
-        ):
+        if is_authentication_error(ex):
             raise AuthenticationError(f"Failed to authenticate {username}") from ex
         raise
     return aws